scripts: improve taco_to_burrito.py
slight improvements for taco_to_burrito.py I made while using it to
convert Teh's Trails.

- add rudimentary support for command line arguments (src + dest). note:
this is not good python code, but as I understand it, converting via
python is only supposed to be a temporary solution.
- also create DEST_DIR/data directory to prevent an error-on-first-write
- support Trails-as-MarkerCategory by adding the texture attribute if it
is present in the <MarkerCategory> and using it when generating Trail
data
- add some more allowed attribs for MarkerCategory and POIs

with these changes, I was able to convert Teh's Trails into a working
burrito format, after making some adjustments to the original files
(fixing some file names and removing all but one color of trails in the
xml).

Signed-off-by: Thorben Römer <thorbenroemer@t-online.de>