simplifying error logging
Adding in a couple of variables to rapidxmls document object so they can be parsed immediately when the error message is created. Error messages are now processed immediately and stored internally so there is no risk of the contents of the error message leaving scope and being unallocated before the error message can be printed to the user.