Don't regenerate files that don't change
The code generator deletes and regenerates all files. This is good in general but it causes the cpp compiler to rebuild the file because it thinks it is a newly changed file. This moves the delete logic until after the regeneration, prevents writing to files if their contents dont change, and prevents the delete logic from deleting any files that it should not.