Renaming CanFade to DisablePlayerCutout
This change also adds the necessary logic for parsing inverted bool xml attributes, and changes the necessary files that need to be changed alongside the primary change.

It does not rename the integration tests because we are not 100% sure how we want to organize the integration tests yet and the tests still technically test the can_fade xml attribute.